wc: Fix bug

+++ web-canvas/src/main/kotlin/pl/karol202/uranium/webcanvas/values/Color.kt
@@ -10,7 +10,7 @@ sealed class Color(val asText: String) : FillStyle
 
 		fun rgb(red: Int, green: Int, blue: Int) = RGBColor(red, green, blue)
 
-		fun rgba(red: Int, green: Int, blue: Int, alpha: Int) = RGBAColor(red, green, blue, alpha)
+		fun rgba(red: Int, green: Int, blue: Int, alpha: Double) = RGBAColor(red, green, blue, alpha)
 	}
 
 	data class RawColor(val name: String) : Color(name)
@@ -22,7 +22,7 @@ sealed class Color(val asText: String) : FillStyle
 	data class RGBAColor(val red: Int,
 	                     val green: Int,
 	                     val blue: Int,
-	                     val alpha: Int) : Color("rgb($red,$green,$blue,$alpha)")
+	                     val alpha: Double) : Color("rgb($red,$green,$blue,$alpha)")
 
 	override fun createNativeStyle(context: DrawContext) = asText
 }
